New Script to find App Protection policy unlicensed users
In Azure Portal -> Intune -> "Client Apps" -> "App protection status"  there is a dashboard element that reports "Unprotected and unlicensed" users, however won't say who is. 
Due to that I've created a new script to find those users which are Assigned to an App Protection policy although they don't have an Intune license.
Currently doesn't support recursive search but I plan to add that in the future.